+++ application/modules/setup/models/Import.php
@@ -53,6 +53,7 @@ class Setup_Model_Import extends Zend_Db_Table_Abstract
         'user_site_form_layout',
         'user_site_form_setting',
         'user_site_history',
+        'user_site_html_styling',
         'user_site_image_library',
         'user_site_image_library_category',
         'user_site_image_library_link',
@@ -74,10 +75,10 @@ class Setup_Model_Import extends Zend_Db_Table_Abstract
         'user_site_page_structure_column_responsive',
         'user_site_page_structure_content',
         'user_site_page_structure_row',
+        'user_site_page_styling',
         'user_site_page_styling_column_background_color',
         'user_site_page_styling_content_item_background_color',
         'user_site_page_styling_content_item_typography',
-        'user_site_page_styling_page_background_color',
         'user_site_page_styling_row_background_color'
     );