+++ library/Dlayer/Model/View/Page/Styling.php
@@ -77,7 +77,20 @@ class Dlayer_Model_View_Page_Styling extends Zend_Db_Table_Abstract
     }
 
     /**
-     * Return the page styles, grouped by styling group and then page id, contains all the data for the selected
+     * Return the content container styles, grouped by styling group and then page id, contains all the data for the
+     * selected page content container
+     *
+     * @return array
+     */
+    public function contentContainer()
+    {
+        return array(
+            'background_color' => $this->contentContainerBackgroundColors()
+        );
+    }
+
+    /**
+     * Return the HTML styles, grouped by styling group and then page id, contains all the data for the selected
      * page
      *
      * @return array
@@ -85,7 +98,20 @@ class Dlayer_Model_View_Page_Styling extends Zend_Db_Table_Abstract
     public function page()
     {
         return array(
-            'background_color' => $this->pageBackgroundColors()
+            'background_color' => $this->contentContainerBackgroundColors()
+        );
+    }
+
+    /**
+     * Return the HTML styles, grouped by styling group and then page id, contains all the data for the selected
+     * page
+     *
+     * @return array
+     */
+    public function html()
+    {
+        return array(
+            'background_color' => $this->htmlBackgroundColor()
         );
     }
 
@@ -229,27 +255,61 @@ class Dlayer_Model_View_Page_Styling extends Zend_Db_Table_Abstract
     }
 
     /**
-     * Fetch the background color styles for a page
+     * Fetch the background color styles for the content container
      *
      * @return array
      */
-    private function pageBackgroundColors()
+    private function contentContainerBackgroundColors()
     {
-        $sql = "SELECT page_id, background_color 
-                FROM user_site_page_styling_page_background_color 
-                WHERE site_id = :site_id 
-                AND page_id = :page_id";
+        $sql = "SELECT 
+                    `page_id`, 
+                    `value` 
+                FROM 
+                    `user_site_page_styling` 
+                WHERE 
+                    `site_id` = :site_id AND 
+                    `page_id` = :page_id AND 
+                    `attribute` = :attribute";
         $stmt = $this->_db->prepare($sql);
-        $stmt->bindValue(':site_id', $this->site_id);
-        $stmt->bindValue(':page_id', $this->page_id);
+        $stmt->bindValue(':site_id', $this->site_id, PDO::PARAM_INT);
+        $stmt->bindValue(':page_id', $this->page_id, PDO::PARAM_INT);
+        $stmt->bindValue(':attribute', 'background-color', PDO::PARAM_STR);
         $stmt->execute();
 
         $styles = array();
 
         foreach ($stmt->fetchAll() as $row) {
-            $styles[intval($row['page_id'])] = $row['background_color'];
+            $styles[intval($row['page_id'])] = $row['value'];
         }
 
         return $styles;
+    }
+
+    /**
+     * Fetch the background color styles for HTML/page
+     *
+     * @return string|false
+     */
+    private function htmlBackgroundColor()
+    {
+        $sql = "SELECT 
+                    `value` 
+                FROM 
+                    `user_site_html_styling` 
+                WHERE 
+                    `site_id` = :site_id AND 
+                    `attribute` = :attribute";
+        $stmt = $this->_db->prepare($sql);
+        $stmt->bindValue(':site_id', $this->site_id, PDO::PARAM_INT);
+        $stmt->bindValue(':attribute', 'background-color', PDO::PARAM_STR);
+        $stmt->execute();
+
+        $result = $stmt->fetch();
+
+        if ($result != false) {
+            return $result['value'];
+        } else {
+            return false;
+        }
     }
 }