+++ private/database/tables/keys/user_site_html_styling.sql
@@ -1,0 +1,3 @@
+
+ALTER TABLE `dlayer`.`user_site_html_styling`
+    ADD CONSTRAINT `user_site_html_styling_fk1` FOREIGN KEY (`site_id`) REFERENCES `user_site` (`id`);